+++ helpers_test.go
@@ -16,3 +16,14 @@ func recoverPanic(f func()) (recovered interface{}) {
 	f()
 	return nil
 }
+
+// like [errors.Join] in the standard library. But this library supports go
+// 1.19, which doesn't have errors.Join.
+func errorsJoin(errs ...error) error {
+	for _, err := range errs {
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}