+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support to setup an `Inspector`, and `InspectorClient` to receive output from `console` messages in JS code.
+- Add support for `FunctionTemplate.Inherit` to setup prototype inheritance.
 
 ### Changed